- merge the exception handlers together - removes unneeded bad request exception
