--- src/main/java/uk/gov/hmcts/reform/notificationservice/exception/BadRequestException.java
@@ -1,17 +1,0 @@
-package uk.gov.hmcts.reform.notificationservice.exception;
-
-import java.io.Serial;
-
-public class BadRequestException extends RuntimeException{
-
-    @Serial
-    private static final long serialVersionUID = 7579941826346533850L;
-
-    public BadRequestException(Integer notificationId) {
-        super("A BadRequestException was received from supplier. Notification has been marked as a fail. Notification ID: " + notificationId);
-    }
-
-    public BadRequestException(Exception exception) {
-        super(exception);
-    }
-}
